move PageSelector
