+++ game/frontend/src/components/PageSelector.vue
